pyproject.toml
[project]
name = "mnemonic"
version = "0.1.0"
description = "Windows EXE to Android APK converter CLI tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "chardet>=5.0.0",
    "pillow>=12.1.0",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.28.1",
]

[project.scripts]
mnemonic = "mnemonic.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mnemonic"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "e2e: marks tests as end-to-end tests (deselect with '-m \"not e2e\"')",
]

[tool.mypy]
python_version = "3.12"
strict = true

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.12.20250915",
]
